requirements.txt
python-telegram-bot
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
openai
apscheduler
python-dotenv
flask
gunicorn
